chore(refactor): refactor unit test
