+++ programs/Anchor.toml
@@ -5,7 +5,7 @@ resolution = true
 skip-lint = false
 
 [programs.localnet]
-programs = "6YpGMWKrqbbsspCbUgXFH1x7RiSnoaXuS3FsG4GWfTdf"
+programs = "89S48feUon6ffgtLzsnqoBVwdb1mxT4rmhRR5WnYefpA"
 
 [registry]
 url = "https://api.apr.dev"
